Refactoring

+++ src/level/Level.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include "Level.h"
 
 void Level::draw(sf::RenderTarget &target, TileSet& tileset)
@@ -10,7 +9,7 @@ void Level::draw(sf::RenderTarget &target, TileSet& tileset)
             LevelTile t = tiles[x][y];
             auto id = t.tile_id;
             auto sprite = tileset.make_sprite(id);
-            sprite.setPosition(x * TILE_X_SIZE, y * TILE_Y_SIZE);
+            sprite.setPosition(static_cast<float>(x * TILE_X_SIZE), static_cast<float>(y * TILE_Y_SIZE));
             target.draw(sprite);
         }
     }
